Renames the albums from album.
