--- src/albums.js
@@ -1,8 +1,0 @@
-/* global fetch:false */
-
-export const getAlbum = id => (
-  fetch(`https://api.spotify.com/v1/albums/${id}`)
-    .then(data => data.json())
-);
-
-export const getAlbumTracks = () => {};
